requirements.txt

---

# requirements.txt
```text
Flask==2.2.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
SQLAlchemy==2.0.20
PyMySQL==1.1.0
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
Werkzeug==2.2.3
xhtml2pdf==0.2.7
alembic==1.11.1
pytz==2024.1
gunicorn==21.2.0
